validation is defined in handlererrors and executed in handler
