+++ internal/handler/wire.go
@@ -15,11 +15,14 @@
 package handler
 
 import (
+	"errors"
 	"fmt"
+	"math"
 
 	"github.com/FerretDB/wire"
 
 	"github.com/FerretDB/FerretDB/internal/bson"
+	"github.com/FerretDB/FerretDB/internal/handler/handlererrors"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/util/must"
@@ -30,7 +33,7 @@ func NewOpMsg(doc *types.Document) (*wire.OpMsg, error) {
 	if err := validateValue(doc); err != nil {
 		doc.Remove("lsid") // to simplify error message
 
-		return nil, newValidationError(fmt.Errorf("wire.OpMsg.Document: validation failed for %v with: %v",
+		return nil, handlererrors.NewValidationError(fmt.Errorf("wire.OpMsg.Document: validation failed for %v with: %v",
 			types.FormatAnyValue(doc),
 			err,
 		))
@@ -78,11 +81,38 @@ func OpMsgDocument(msg *wire.OpMsg) (*types.Document, error) {
 	if err = validateValue(res); err != nil {
 		res.Remove("lsid") // to simplify error message
 
-		return nil, newValidationError(fmt.Errorf("wire.OpMsg.Document: validation failed for %v with: %v",
+		return nil, handlererrors.NewValidationError(fmt.Errorf("wire.OpMsg.Document: validation failed for %v with: %v",
 			types.FormatAnyValue(res),
 			err,
 		))
 	}
 
 	return res, nil
+}
+
+// validateValue checks given value and returns error if unsupported value was encountered.
+func validateValue(v any) error {
+	switch v := v.(type) {
+	case *types.Document:
+		for _, v := range v.Values() {
+			if err := validateValue(v); err != nil {
+				return err
+			}
+		}
+
+	case *types.Array:
+		for i := 0; i < v.Len(); i++ {
+			v := must.NotFail(v.Get(i))
+			if err := validateValue(v); err != nil {
+				return err
+			}
+		}
+
+	case float64:
+		if math.IsNaN(v) {
+			return errors.New("NaN is not supported")
+		}
+	}
+
+	return nil
 }
